requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.6
pydantic==1.10.7
requests==2.29.0
sniffio==1.3.0
SQLAlchemy==2.0.11
starlette==0.26.1
typing-extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
